Created a new version with OSX support
Big shift in the design.
+ Dropped GPG requirement. Opted for a large binary key stored in the
home dir of the user account.
+ Created an OSX install script that runs (command line only) to
duplicate the setup, but in an HFS+ encrypted volume.
+ The OSX version requires a short script to be run each time, but it
is automatically created by the installer. The encryption is secured by
a passphrase that must be memorized.

Next steps:
+ Adjust the setup script to allow for central creation of the USBkey,
but a short installer script to be run by the unprivileged user.
+ Must use 'shred' to delete the keys!