New workflow
Made some big adjustments to the workflow so that we can create the keys
locally, and allow the completion to finish on the target workstation
(and OSX device when requested). For this to work I needed to adjust
some of the order of operations, to account for the appropriate
difference in user accounts / environments.

Untested at this point.