Minor improvements
Added an insertion hint file that will prevent the control script from
locking the workstation if the wrong USBkey were inserted (or something
failed) at insertion time. This could always be deleted accidentally (or
maliciously) but since the SSH keys have a lifetime on them, its not a
large risk.
Also added an error check to make sure we don't execute if the target
user isn't identified or X isn't running.