Added support for D-Bus based screen locking
When we moved from xscreensaver to light-locker, we noticed the
automatic locking feature of the USBkey stopped working. This was due to
the implementation for light-locker using D-Bus messages to trigger a
lock as opposed to simply the DISPLAY setting. It was pretty easy to
incorporate an appropriate check for the active D-Bus for the target
user, and the added benefit was that I learned how to remove the
hardcoded DISPLAY value for the other lockers (closing Issue #2).